upgrade changes and more unit test
